Sim eval: generic SimRolloutEval + per-algo sim_init_state/sim_predict_step
The closed-loop env-rollout logic was previously HNet-specific. Refactored
so each algo owns its own inference state + step semantics and the eval
class is fully algo-agnostic.

**New: egomimic/eval/eval_sim.py — SimRolloutEval**:
- Owns env reset/step/render, replay-vs-random init, coverage/success
  metrics, video buffering.
- Per-embodiment obs formatter table (``_OBS_FORMATTERS``) handles the
  env -> model obs shape conversion.
- Calls ``algo.sim_init_state(...)`` once per episode, then
  ``algo.sim_predict_step(state, obs_norm, t, emb_id)`` per step.

**New: HNet.sim_init_state / HNet.sim_predict_step** (egomimic/algo/hnet.py):
- ``sim_init_state``: allocates the AR inference cache, primes the
  initial input token. Handles both HNetPolicy and FlatFusedPolicy.
- ``sim_predict_step``: encodes single-frame obs, runs one transformer
  step, returns normalized action ``(B, 1, action_dim)``. Mutates state.

These two methods are the **shared inference API** for AR-style closed-loop
rollout. ``robot/rollout.py`` and other consumers can call them directly
without going through SimRolloutEval — the eval class is just one consumer.

**eval_hnet_sim.py** is now a back-compat shim that re-exports
``SimRolloutEval`` as ``HNetSimEval``.

HPT's equivalents (chunk-based ``sim_predict_step``) land in a follow-up
commit alongside the new HPT model config + training run.
